* Zahl der konfigurierten Geräte wird überprüft.  * Einfaches Licht wird direkt in seriellen Puffer übertragen.  * Controller mit ID 11 simuliert - kein Problem ?!?  * int in uint8_t übertragen.
git-svn-id: svn+ssh://svn.code.sf.net/p/mrw/code/trunk@117 0b73ed94-99fe-4531-ab00-e2d4daa7083f

+++ mc/config.c
@@ -83,7 +83,8 @@ void init_light_signal(
 {
 	dvc->unit_type = type;
 	dvc->unit.u_signal.count = count;
-	for (int i = 0;i < count;i++)
+
+	for (uint8_t i = 0;i < count;i++)
 	{
 		dvc->unit.u_signal.byte[i] = buffer[i] >> 3;
 		dvc->unit.u_signal.bit[i]  = buffer[i] & 0x7;
